I've finished tests for the models. Unfortunately I'm using should syntax rather than expect. Something to change later.
I forgot about one of the associations that are currently there for the order model. Just added a test for that.

I've added tests for option model. basic properties and associations, still want to come back to this though. I'm not sure if I'm using the right test for the has_many_through.

added basic tests for lineitem model.

I've added basic tests for category model.